@@ -13,14 +13,14 @@ var (
 	requestCounter = metrics.NewCounter("iost_rpc_request", []string{"method"})
 )
 
-func metricsUnaryMiddleware(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error) {
+func metricsUnaryMiddleware(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
 	i := strings.LastIndex(info.FullMethod, "/")
 	ilog.Debugf("receive rpc request: %s, request: %v", info.FullMethod[i+1:], req)
 	requestCounter.Add(1, map[string]string{"method": info.FullMethod[i+1:]})
 	return handler(ctx, req)
 }
 
-func metricsStreamMiddleware(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
+func metricsStreamMiddleware(srv any, ss grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
 	i := strings.LastIndex(info.FullMethod, "/")
 	ilog.Debugf("receive rpc stream: %s", info.FullMethod[i+1:])
 	requestCounter.Add(1, map[string]string{"method": info.FullMethod[i+1:]})
