@@ -77,14 +77,14 @@ func (e *VM) setHost(host *host.Host) {
 	e.sandbox.SetHost(host)
 }
 
-func (e *VM) setContract(contract *contract.Contract, api string, args []interface{}) (string, error) {
+func (e *VM) setContract(contract *contract.Contract, api string, args []any) (string, error) {
 	return e.sandbox.Prepare(contract, api, args)
 }
 
-func (e *VM) execute(code string) (rtn []interface{}, cost contract.Cost, err error) {
+func (e *VM) execute(code string) (rtn []any, cost contract.Cost, err error) {
 	rs, gasUsed, err := e.sandbox.Execute(code)
 	gasCost := contract.NewCost(0, 0, gasUsed)
-	return []interface{}{rs}, gasCost, err
+	return []any{rs}, gasCost, err
 }
 
 // EnsureFlags make sandbox flag is same with input
