@@ -12,6 +12,6 @@ type VM interface {
 	Init() error
 	Validate(contract *contract.Contract) error
 	Compile(contract *contract.Contract) (string, error)
-	LoadAndCall(host *host.Host, contract *contract.Contract, api string, args ...interface{}) (rtn []interface{}, cost contract.Cost, err error)
+	LoadAndCall(host *host.Host, contract *contract.Contract, api string, args ...any) (rtn []any, cost contract.Cost, err error)
 	Release()
 }
